Pick unicode/ascii agnostic tree characters
These are needed to workaround the constraint that Python 2 can only
place `bytes` in `__repr__` results and Python 3 can only place
`unicode` in `__repr__` results. As such, we need to chose characters
that will work well in both ASCII and UTF-8. Hence we change the symbols
that `asciitree` uses to be a set that match this criteria.
Unfortunately they cannot be imported from `asciitree` directly as they
force all of their characters to UTF-8.