Use `os.renames` in `DirectoryStore`'s `rename`
Instead of using `os.rename` and creating missing directories before
when renaming content in a `DirectoryStore`, just use `os.rename` to
handle this move for us. This will automatically create missing
directories for us in the destination path so we don't have to. Also it
prunes away any empty directories in the old path, which we were not
doing before. IOW this fixes a bug as well.