Ensure `DictStore` contains only `bytes` (#350)
* Ensure `DictStore` gets `bytes` coercible data

As the spec requires that the data in a store be a sequence of `bytes`,
make sure that non-`DictStore` input meets this requirement when setting
values. This effectively ensures that other `DictStore` meet this
requirement as well. So we don't need to go through and check their
values too.

* Drop undefined buffer size case from `DictStore`

As everything in `DictStore` must either be another `DictStore` or
`bytes`, there shouldn't be any cases where the size is undefined nor
cases that this exception should need handling. Given this go ahead and
drop the special casing for unknown sizes in `DictStore`.

* Drop `test_getsize_ext`

While this test case does test a useful subset of the `getsize` API, the
contents being added to the store here are non-conforming to our
expectations of store contents. Namely the store should only contain
values that are an "arbitrary sequence of bytes", which this test case
is not.

* Test `getsize` with an unknown size case

This creates a non-conforming store to make sure that `getsize` handles
its contents in the expected way. Namely that it returns `-1`.

* Note `DictStore` only contains `bytes` now        [ci skip]

* Check for `TypeError` for non-buffer objects

Add a test to ensure that a non-buffer supporting object when stored
into a valid store, will raise a `TypeError` instead of storing it.
Disable this checking for generic `MappingStore`s (e.g. `dict`) as they
do not perform this sort of checking on the data they accept as values.

* Check that `DictStore` coerces all data to `bytes`

Provide a simple test for `DictStore` to ensure that non-`bytes` is
coerced to `bytes` before storing it and is retrieved as `bytes`.

* Disallow mutation of the internal `DictStore`

Make sure that users are only able to add data to the `DictStore`.
Disallow the storing of a nested `DictStore` though.