Always use `tuple`s for multidimensional indexing (#376)
* Always use `tuple`s for multidimensional indexing

NumPy 1.15.0 deprecates the use of sequences in `ndarray` selection for
sequences other than `tuple`s. We have a few cases where we are using
`list`s instead as they are mutable. Thus are easier to build up and
change. However passing `list`s into `ndarray` selection will cause this
`FutureWarning` to be emitted (and will eventually result in an error).

To fix this simply, just convert these `list`s to `tuple`s after we are
done changing their contents, but before we use them with `ndarray`s.
This works just as well on older versions of NumPy. Avoids the
`FutureWarning` in NumPy 1.15.0+. Not to mention this will continue to
work on future versions of NumPy, which may change this warning into an
error.

* Note use of `tuple`s for slicing NumPy `ndarrays`