Collect JSON utility functions (#430)
* Move `ensure_text_type` to `zarr.util`

As `ensure_text_type` is a more general utility function, go ahead and
move it over to `zarr.util` and drop it from `zarr.meta`. Though make
sure to import it into `zarr.meta` so that it can still be used there.

* Move `json_dumps` to `zarr.util`

As `json_dumps` has broader use than just for metdata (e.g. attributes
would also benefit from it), this moves `json_dumps` to `zarr.util`.

* Move `json_loads` to `zarr.util`

As `json_loads` is more generally useful than just parsing metadata,
move it over to `zarr.util`. Though ensure it is still imported in
`zarr.meta`.

* Use `json_loads` in `ConsolidatedMetadataStore`

Should ensure that the `ConsolidatedMetadataStore` is able to handle
loading JSON data in a standard way.

* Use `json_dumps` to serialize attributes

* Handle `encode` within `json_dumps`

As every use of `json_dumps` is followed by a call to `encode` to
convert the string to bytes, go ahead and incorporate this step in
`json_dumps`. Should make it a bit easier to use.

* Drop `encode` from `json_dumps` output

As the `encode` step is now part of `json_dumps`, remove that step
when using the `json_dumps` function.

* Drop unneeded intermediates in `json_dumps` usage

After moving `encode` within `json_dumps`, there were
intermediate variables left over that weren't really needed. This
drops them.

* Update release note on JSON reading and writing