Ensure contiguous data using `astype` (#513)
Previously we were calling `ascontiguousarray` and `asfortranarray`
based on whether we wanted a C or Fortran ordered array. However we can
simplify this by using `astype` to capture both cases. Additionally
those functions ensure the array is at least 1-D, but that doesn't seem
to matter for this use case since the scalar case is handled in a
different branch. So we leave that part out.