Add guard as to version number;
It seem like setuptools scm can screw up and this should ensure we catch
that in CI before making the wheel. We want to make sure to have a
runtime check for the version number as well as a check just after
building the wheel.

This seem to be due to `build` latest version, so i've pinned it for
now.