Delete unused files (#1251)
* Delete build.cmd

This file was leftover from before numcodecs was split from zarr-python.
It was used to aid in compiling the Cython extensions on Windows.
However zarr-python pure Python. So there is not a need to keep this
file. It also doesn't appear to be used here. So drop it.

* Update release.rst

* Delete unused `.gitmodules` as well

This was for pointing to Blosc. Again not needed after the
zarr-python/numcodecs split.

* Drop `release.txt`

This is unused. Releases are handled through GitHub Actions. One should
look there if attempting to reproduce outside of that infrastructure.

* Drop unused `windows_conda_dev.txt`

* Generalize release entry